Create wirebase class
Uses existing BlockBluectricCable as a template.

Code was copied from BlockBluectricCable and updated to remove
Bluelectric specifics and provide means to perform the same actions.

The goal is to have a single base class for the wire types ( Bluelectric, RedAlloy, ... ).
These wire types are so similar that having most of the logic in the base class will
simplify the wire logic, and provide a single place to update their code.